refactor: update entity IDs to Long in DTOs and model classes

+++ src/main/java/org/ifpb/dto/consulta/ConsultaRequestDTO.java
@@ -18,8 +18,8 @@ import java.time.LocalDateTime;
  * */
 public class ConsultaRequestDTO {
 
-    private int idPaciente;
-    private int idMedico;
+    private Long idPaciente;
+    private Long idMedico;
     private LocalDateTime data;
 
 }
